Tidy up target index calculation
This commit updates determineTargetIndex() to be more consistent
with other implementations.

The new implementation is roughly the same as before, but hopefully
clearer to understand. The changes mean that we can now use a round,
rather than a truncate. This which means that flings which travel
most of the distance will now target the 'next' page. Previously they
would truncate down towards 0.