Fix float inaccuracies at high values
When determining the target index to fling to, we are currently
including the current index into a float.

This means that if the current scroll index is a large number (millions),
the accuracy of the decimals below that is greatly impacted, enough so to
affect rounding and the calculated scroll index.

This PR fixes that in 2 ways. First, we use double calculations for determining
the index offset. Secondly, we keep the current index out of the double.

Fixes #13