Add `snapIndex` lambda parameter
This block which returns the index which the apps wishes to snap to.
The block is provided with the SnapperLayoutInfo and the index which Snapper
has determined is the correct target index. Callers can override this value as
they see fit.

A common use case could be rounding up/down to achieve groupings of items, which was
the intention of #12