Update to Compose 1.2.0 (#31)
* Update to Compose 1.2.0 DO NOT MERGE

This PR updates Snapper to build against Compose 1.2.0. We do not
want to merge this until Compose 1.2.0 final is out though.

Primarily this is so that we can use the new `afterContentPadding`
exposed by `LazyListLayoutInfo`. This means that we developers no
longer need to manually pass in that value 🎉.

I've deprecated the relevant functions with the necessary `ReplaceWith`s.

* Update dependencies

* Fix metalava API file